refactor: split expansion helpers into modules
