+++ tests/run_var_tests.sh
@@ -45,6 +45,9 @@ test_export_p_listing.expect
 test_export_n_unexport.expect
 test_readonly_p.expect
 test_set_list.expect
+test_field_split_module.expect
+test_param_expand_module.expect
+test_quote_utils_module.expect
 "
 for test in $tests; do
     echo "Running $test"
