+++ tests/test_field_split_module.expect
@@ -1,0 +1,27 @@
+#!/usr/bin/env expect
+set timeout 5
+spawn [file dirname [info script]]/../build/vush
+expect {
+    "vush> " {}
+    timeout { send_user "prompt timeout\n"; exit 1 }
+}
+send "IFS=:\r"
+expect {
+    "vush> " {}
+    timeout { send_user "set IFS failed\n"; exit 1 }
+}
+send "set -- a:b::c:\r"
+expect {
+    "vush> " {}
+    timeout { send_user "set args failed\n"; exit 1 }
+}
+send "echo \$#\r"
+expect {
+    -re "\r\n5\r\nvush> " {}
+    timeout { send_user "count failed\n"; exit 1 }
+}
+send "exit\r"
+expect {
+    eof {}
+    timeout { send_user "eof timeout\n"; exit 1 }
+}
